Rename middleware to fixtures, remove unused code
