+++ e2e/fixtures/wallet_balance.selectors.ts
