+++ e2e/fixtures/wallet_balance.ts
