+++ e2e/fixtures/wallet_connect.selectors.ts
