+++ e2e/fixtures/wallet_connect.ts
